+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Bugfixes:
 * Improve support of `--help *-format` and several bugs are found and fixed
 * Don't incorrectly print `No active sound devices found` when using a non-controllable sound device (Sound, macOS)
 
+Logo:
+* Add Chimera Linux
+
 # 2.1.0
 
 This release introduces a new output format: JSON result